load-data-file-service/requirements.txt
fastapi~=0.111.0
dnspython~=2.6.1
idna~=3.7
httpcore~=1.0.5
httpx~=0.27.0
anyio~=4.4.0
h11~=0.14.0
pip~=23.2.1
wheel~=0.41.2
Jinja2~=3.1.4
typing_extensions~=4.12.2
rich~=13.7.1
Pygments~=2.18.0
PyYAML~=6.0.1
exceptiongroup~=1.2.1
sniffio~=1.3.1
click~=8.1.7
certifi~=2024.6.2
mdurl~=0.1.2
typer~=0.12.3
shellingham~=1.5.4
setuptools~=68.2.0
MarkupSafe~=2.1.5
orjson~=3.10.5
starlette~=0.37.2
email_validator~=2.2.0
pydantic~=2.7.4
pydantic_core~=2.18.4
websockets~=12.0
ujson~=5.10.0
uvicorn~=0.30.1
httptools~=0.6.1
watchfiles~=0.22.0
python-multipart~=0.0.9
requests~=2.32.3
python-dotenv~=1.0.1
pyodbc
pandas~=2.2.2
numpy~=2.0.0